feat: test refactoring
